+++ src/Trarizon.Library/Collections/TraCollection/List.cs
@@ -2,8 +2,23 @@
 using System.Runtime.InteropServices;
 
 namespace Trarizon.Library.Collections;
-public static partial class TraList
+public static partial class TraCollection
 {
+#if NETSTANDARD
+
+    /// <remarks>
+    /// As CollectionsMarshal doesnt exists on .NET Standard 2.0, this use a very tricky way
+    /// to get the underlying array. Actually I'm not sure if this works correctly in all runtime...
+    /// (at leat it works on Unity
+    /// </remarks>
+    public static Span<T> AsSpan<T>(this List<T> list) 
+        => Utils<T>.GetUnderlyingArray(list).AsSpan(..list.Count);
+
+#endif
+
+    public static T[] GetUnderlyingArray<T>(List<T> list)
+        => Utils<T>.GetUnderlyingArray(list);
+
     public static void RemoveAt<T>(this List<T> list, Index index)
         => list.RemoveAt(index.GetOffset(list.Count));
 
@@ -18,13 +33,7 @@ public static partial class TraList
         list.AsSpan().MoveTo(fromIndex, toIndex);
         // If we modified in Span, the field _version of List<> won't update, so here we manually
         // use SetCount to update _version;
-#if NET9_0_OR_GREATER
         Utils<T>.GetVersion(list)++;
-#elif NETSTANDARD
-        list[0] = list[0];
-#else
-        CollectionsMarshal.SetCount(list, list.Count);
-#endif
     }
 
     public static void MoveTo<T>(this List<T> list, Index fromIndex, Index toIndex)
@@ -37,13 +46,7 @@ public static partial class TraList
     {
         list.AsSpan().MoveTo(fromIndex, toIndex, moveCount);
         // See MoveTo(List<>, int, int) for explanation
-#if NET9_0_OR_GREATER
         Utils<T>.GetVersion(list)++;
-#elif NETSTANDARD
-        list[0] = list[0];
-#else
-        CollectionsMarshal.SetCount(list, list.Count);
-#endif
     }
 
     public static void MoveTo<T>(this List<T> list, Range fromRange, Index toIndex)